backend (migrated)/requirements.txt
firecrawl-py>=1.4.4
beautifulsoup4==4.12.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
